# tests/data/openapi/json/openapi.json
{"openapi":"3.0.2","info":{"title":"Fast API","version":"0.1.0"},"paths":{"/schema":{"post":{"summary":"New Schema","description":"New Schema.","operationId":"new_schema_schema_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpenAPISchemaV3"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/mormo__schema__openapi_v3__SaveDBResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/schema/{id}":{"get":{"summary":"Get Schema","operationId":"get_schema_schema__id__get","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpenAPISchemaV3"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/schema/{id}/to_postman":{"post":{"summary":"Schema To Postman","description":"Convert schema to Collection.","operationId":"schema_to_postman_schema__id__to_postman_post","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpenAPISchemaToPostmanRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/mormo__schema__postman_collection_v2__SaveDBResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/postman/{id}":{"get":{"summary":"Get Postman","description":"Get Postman Collection by ID.","operationId":"get_postman_postman__id__get","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/postman/{id}/test":{"get":{"summary":"Run Postman Test","description":"Create a new test run from a postman collection.","operationId":"run_postman_test_postman__id__test_get","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"},{"required":true,"schema":{"title":"Host","type":"string"},"name":"host","in":"query"},{"required":false,"schema":{"title":"Verbose","type":"boolean","default":false},"name":"verbose","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/run/test/from_schema":{"post":{"summary":"Run Test Run From Schema","description":"Create a new test run from OpenAPI Schema.","operationId":"run_test_run_from_schema_run_test_from_schema_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpenAPISchemaToPostmanRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Auth":{"title":"Auth","required":["type"],"type":"object","properties":{"type":{"title":"Type","type":"string"}}},"Collection":{"title":"Collection","required":["item","info"],"type":"object","properties":{"item":{"title":"Item","type":"array","items":{"anyOf":[{"$ref":"#/components/schemas/Item"},{"$ref":"#/components/schemas/Folder"}]}},"event":{"title":"Event","type":"array","items":{}},"variable":{"title":"Variable","type":"array","items":{"$ref":"#/components/schemas/Variable"}},"info":{"$ref":"#/components/schemas/mormo__schema__postman_collection_v2__Info"}}},"Contact":{"title":"Contact","type":"object","properties":{"name":{"title":"Name","type":"string"},"url":{"title":"Url","type":"string"},"email":{"title":"Email","type":"string"}}},"Description":{"title":"Description","required":["content","type"],"type":"object","properties":{"content":{"title":"Content","type":"string"},"type":{"title":"Type","type":"string"}}},"Encoding":{"title":"Encoding","type":"object","properties":{"contentType":{"title":"Contenttype","type":"string"},"headers":{"title":"Headers","type":"object","additionalProperties":{"anyOf":[{"$ref":"#/components/schemas/Reference"},{"type":"object"}]}},"style":{"title":"Style","type":"string"},"explode":{"title":"Explode","type":"boolean"},"allowReserved":{"title":"Allowreserved","type":"boolean"}}},"Event":{"title":"Event","required":["listen","script"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"listen":{"title":"Listen","type":"string"},"disabled":{"title":"Disabled","type":"boolean"},"script":{"$ref":"#/components/schemas/Script"}}},"Expect":{"title":"Expect","type":"object","properties":{"code":{"title":"Code","anyOf":[{"type":"integer"},{"type":"string"}]},"enabled":{"title":"Enabled","type":"boolean","default":true},"response_time":{"title":"Response Time","type":"integer","default":200},"headers":{"title":"Headers","type":"object","additionalProperties":{"type":"string"}},"fake_data":{"title":"Fake Data","type":"boolean","default":true},"enabled_tests":{"title":"Enabled Tests","type":"array","items":{"enum":["schema_validation","code","content_type","response_time"]},"default":["schema_validation","code","content_type","response_time"]}}},"ExternalDocs":{"title":"ExternalDocs","required":["url"],"type":"object","properties":{"description":{"title":"Description","type":"string"},"url":{"title":"Url","type":"string"}}},"Folder":{"title":"Folder","required":["id","name","item","event"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"item":{"title":"Item","anyOf":[{"type":"array","items":{"$ref":"#/components/schemas/Item"}},{"$ref":"#/components/schemas/Item"}]},"event":{"title":"Event","type":"array","items":{}},"variable":{"title":"Variable","type":"array","items":{"$ref":"#/components/schemas/Variable"}}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Item":{"title":"Item","required":["id","name","request","response","event"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"request":{"$ref":"#/components/schemas/Request"},"response":{"title":"Response","type":"array","items":{"$ref":"#/components/schemas/mormo__schema__postman_collection_v2__Response"}},"event":{"title":"Event","type":"array","items":{"$ref":"#/components/schemas/Event"}}}},"License":{"title":"License","required":["name"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"url":{"title":"Url","type":"string"}}},"MediaType":{"title":"MediaType","type":"object","properties":{"schema":{"title":"Schema","anyOf":[{"$ref":"#/components/schemas/Reference"},{"$ref":"#/components/schemas/SchemaObject"}]}}},"NewmanResult":{"title":"NewmanResult","required":["stdout","code"],"type":"object","properties":{"stderr":{"title":"Stderr","type":"string"},"stdout":{"title":"Stdout","type":"string"},"json":{"title":"Json","type":"object"},"code":{"title":"Code","type":"integer"}}},"OpenAPISchemaToPostmanRequest":{"title":"OpenAPISchemaToPostmanRequest","type":"object","properties":{"schema":{"$ref":"#/components/schemas/OpenAPISchemaV3"},"path":{"title":"Path","type":"string"},"host":{"title":"Host","type":"string"},"test_config":{"title":"Test Config","type":"object","additionalProperties":{"$ref":"#/components/schemas/TestConfig"}},"test_data_file":{"title":"Test Data File","type":"string"},"test_data":{"title":"Test Data","type":"array","items":{"$ref":"#/components/schemas/TestData"}},"test_scripts":{"title":"Test Scripts","type":"object","additionalProperties":{"$ref":"#/components/schemas/Script"}},"prerequest_scripts":{"title":"Prerequest Scripts","type":"object","additionalProperties":{"$ref":"#/components/schemas/Script"}},"collection_test_scripts":{"title":"Collection Test Scripts","type":"array","items":{"$ref":"#/components/schemas/Script"}},"collection_prerequest_scripts":{"title":"Collection Prerequest Scripts","type":"array","items":{"$ref":"#/components/schemas/Script"}},"collection_global_variables":{"title":"Collection Global Variables","type":"array","items":{"$ref":"#/components/schemas/Variable"}},"expect":{"title":"Expect","type":"object","additionalProperties":{"$ref":"#/components/schemas/Expect"}},"verbose":{"title":"Verbose","type":"boolean","default":false}}},"OpenAPISchemaV3":{"title":"OpenAPISchemaV3","required":["openapi","paths"],"type":"object","properties":{"openapi":{"title":"Openapi","type":"string"},"paths":{"title":"Paths","type":"object","additionalProperties":{"$ref":"#/components/schemas/Path"}},"info":{"$ref":"#/components/schemas/mormo__schema__openapi_v3__Info"},"servers":{"title":"Servers","type":"array","items":{"$ref":"#/components/schemas/Server"}},"components":{"title":"Components","type":"object"},"security":{"title":"Security","type":"array","items":{"type":"object","additionalProperties":{"type":"array","items":{}}}},"externalDocs":{"$ref":"#/components/schemas/ExternalDocs"}}},"Operation":{"title":"Operation","required":["responses"],"type":"object","properties":{"operationId":{"title":"Operationid","type":"string"},"summary":{"title":"Summary","type":"string"},"responses":{"title":"Responses","type":"object","additionalProperties":{"anyOf":[{"$ref":"#/components/schemas/mormo__schema__openapi_v3__Response"},{"$ref":"#/components/schemas/Reference"}]}},"parameters":{"title":"Parameters","type":"array","items":{"anyOf":[{"$ref":"#/components/schemas/Parameter"},{"$ref":"#/components/schemas/Reference"}]}},"tags":{"title":"Tags","type":"array","items":{"type":"string"}},"externalDocs":{"$ref":"#/components/schemas/ExternalDocs"},"requestBody":{"title":"Requestbody","anyOf":[{"$ref":"#/components/schemas/mormo__schema__openapi_v3__RequestBody"},{"$ref":"#/components/schemas/Reference"}]},"callbacks":{"title":"Callbacks","type":"object","additionalProperties":{"anyOf":[{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"type":"object"}}},{"$ref":"#/components/schemas/Reference"}]}},"deprecated":{"title":"Deprecated","type":"boolean"},"security":{"title":"Security","type":"array","items":{"type":"object","additionalProperties":{"type":"array","items":{}}}},"servers":{"title":"Servers","type":"array","items":{"$ref":"#/components/schemas/Server"}}}},"OriginalRequest":{"title":"OriginalRequest","required":["method","url","body"],"type":"object","properties":{"method":{"title":"Method","type":"string"},"url":{"$ref":"#/components/schemas/Url"},"body":{"title":"Body","type":"object"}}},"Parameter":{"title":"Parameter","required":["name","in","schema"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"in":{"title":"In","enum":["body","requestBody","path","query","header","cookie"]},"description":{"title":"Description","type":"string"},"required":{"title":"Required","type":"boolean"},"schema":{"$ref":"#/components/schemas/ParameterSchema"},"deprecated":{"title":"Deprecated","type":"boolean"},"allowEmptyValue":{"title":"Allowemptyvalue","type":"boolean"},"style":{"title":"Style","type":"string"},"explode":{"title":"Explode","type":"boolean"},"allowReserved":{"title":"Allowreserved","type":"boolean"},"example":{"title":"Example"},"examples":{"title":"Examples","type":"object","additionalProperties":{"anyOf":[{},{"$ref":"#/components/schemas/Reference"}]}},"content":{"title":"Content","type":"object","additionalProperties":{"$ref":"#/components/schemas/MediaType"}}}},"ParameterSchema":{"title":"ParameterSchema","type":"object","properties":{"type":{"title":"Type","type":"string"},"format":{"title":"Format","type":"string"},"title":{"title":"Title","type":"string"},"maximum":{"title":"Maximum","type":"integer"},"minimum":{"title":"Minimum","type":"integer"},"nullable":{"title":"Nullable","type":"boolean"}}},"Path":{"title":"Path","type":"object","properties":{"get":{"$ref":"#/components/schemas/Operation"},"put":{"$ref":"#/components/schemas/Operation"},"post":{"$ref":"#/components/schemas/Operation"},"delete":{"$ref":"#/components/schemas/Operation"},"options":{"$ref":"#/components/schemas/Operation"},"head":{"$ref":"#/components/schemas/Operation"},"patch":{"$ref":"#/components/schemas/Operation"},"trace":{"$ref":"#/components/schemas/Operation"}}},"Properties":{"title":"Properties","type":"object","properties":{"description":{"title":"Description","type":"string"},"type":{"title":"Type","type":"string"},"format":{"title":"Format","type":"string"},"default":{"title":"Default"},"example":{"title":"Example"},"title":{"title":"Title","type":"string"},"multipleOf":{"title":"Multipleof","type":"integer"},"maximum":{"title":"Maximum","type":"integer"},"exclusiveMaximum":{"title":"Exclusivemaximum","type":"boolean"},"minimum":{"title":"Minimum","type":"integer"},"exclusiveMinimum":{"title":"Exclusiveminimum","type":"boolean"},"maxLength":{"title":"Maxlength","type":"integer"},"minLength":{"title":"Minlength","type":"integer"},"pattern":{"title":"Pattern","type":"string"},"additionalItems":{"title":"Additionalitems","anyOf":[{"type":"boolean"},{"type":"object"},{"$ref":"#/components/schemas/Reference"}]},"items":{"title":"Items","anyOf":[{"type":"object"},{"$ref":"#/components/schemas/Reference"}]},"maxItems":{"title":"Maxitems","type":"integer"},"minItems":{"title":"Minitems","type":"integer"},"uniqueItems":{"title":"Uniqueitems","type":"boolean"},"maxProperties":{"title":"Maxproperties","type":"integer"},"minProperties":{"title":"Minproperties","type":"integer"},"required":{"title":"Required","uniqueItems":true,"type":"array","items":{"type":"string"}},"properties":{"title":"Properties","anyOf":[{"type":"object"},{"$ref":"#/components/schemas/Reference"}]},"additionalProperties":{"title":"Additionalproperties","anyOf":[{"type":"boolean"},{"type":"object"},{"$ref":"#/components/schemas/Reference"}]},"enum":{"title":"Enum","uniqueItems":true,"type":"array","items":{}},"allOf":{"title":"Allof","type":"array","items":{"anyOf":[{"type":"object"},{"$ref":"#/components/schemas/Reference"}]}},"oneOf":{"title":"Oneof","type":"array","items":{"anyOf":[{"type":"object"},{"$ref":"#/components/schemas/Reference"}]}},"anyOf":{"title":"Anyof","type":"array","items":{"anyOf":[{"type":"object"},{"$ref":"#/components/schemas/Reference"}]}},"not":{"title":"Not","anyOf":[{"type":"object"},{"$ref":"#/components/schemas/Reference"}]},"nullable":{"title":"Nullable","type":"boolean"},"discriminator":{"title":"Discriminator","type":"object"},"readOnly":{"title":"Readonly","type":"boolean"},"writeOnly":{"title":"Writeonly","type":"boolean"},"xml":{"title":"Xml","type":"object"},"externalDocs":{"$ref":"#/components/schemas/ExternalDocs"},"deprecated":{"title":"Deprecated","type":"boolean"}}},"Reference":{"title":"Reference","required":["$ref"],"type":"object","properties":{"$ref":{"title":"$Ref","type":"string"}}},"Request":{"title":"Request","required":["name","description","method","url","auth","header"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"object"},"method":{"title":"Method","type":"string"},"url":{"$ref":"#/components/schemas/Url"},"auth":{"$ref":"#/components/schemas/Auth"},"header":{"title":"Header","type":"array","items":{}},"body":{"$ref":"#/components/schemas/mormo__schema__postman_collection_v2__RequestBody"}}},"SchemaObject":{"title":"SchemaObject","type":"object","properties":{"description":{"title":"Description","type":"string"},"required":{"title":"Required","type":"array","items":{"type":"string"}},"example":{"title":"Example"},"examples":{"title":"Examples","type":"object","additionalProperties":{"anyOf":[{},{"$ref":"#/components/schemas/Reference"}]}},"encoding":{"title":"Encoding","type":"object","additionalProperties":{"$ref":"#/components/schemas/Encoding"}},"type":{"title":"Type","type":"string"},"format":{"title":"Format","type":"string"},"properties":{"title":"Properties","type":"object","additionalProperties":{"$ref":"#/components/schemas/Properties"}},"allOf":{"title":"Allof","type":"array","items":{"anyOf":[{"type":"object"},{"$ref":"#/components/schemas/Reference"}]}}}},"Script":{"title":"Script","required":["exec"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"type":{"title":"Type","type":"string"},"exec":{"title":"Exec","anyOf":[{"type":"string"},{"type":"array","items":{}}]},"src":{"title":"Src","anyOf":[{"type":"string"},{"$ref":"#/components/schemas/Url"}]},"name":{"title":"Name","type":"string"}}},"Server":{"title":"Server","required":["url"],"type":"object","properties":{"url":{"title":"Url","type":"string"},"description":{"title":"Description","type":"string"},"variables":{"title":"Variables","type":"object","additionalProperties":{"$ref":"#/components/schemas/ServerVariable"}}}},"ServerVariable":{"title":"ServerVariable","required":["default"],"type":"object","properties":{"default":{"title":"Default","type":"string"},"enum":{"title":"Enum","type":"array","items":{"type":"string"}},"description":{"title":"Description","type":"string"}}},"TestConfig":{"title":"TestConfig","type":"object","properties":{"variables":{"title":"Variables","anyOf":[{"type":"object","additionalProperties":{"type":"string"}},{"type":"string"}]},"expect":{"$ref":"#/components/schemas/Expect"},"make_global":{"title":"Make Global","type":"object","additionalProperties":{"type":"string"}},"test":{"title":"Test","type":"array","items":{"type":"string"}},"prerequest":{"title":"Prerequest","type":"array","items":{"type":"string"}},"headers":{"title":"Headers","type":"object","additionalProperties":{"type":"string"}}}},"TestData":{"title":"TestData","required":["route","in_","key"],"type":"object","properties":{"route":{"title":"Route","type":"string"},"in_":{"title":"In ","enum":["body","requestBody","path","query","header","cookie"]},"key":{"title":"Key","type":"string"},"value":{"title":"Value"}}},"TestResult":{"title":"TestResult","required":["result","code","message"],"type":"object","properties":{"result":{"$ref":"#/components/schemas/NewmanResult"},"code":{"title":"Code","type":"integer"},"message":{"title":"Message","type":"string"}}},"Url":{"title":"Url","required":["path","host","query","variable"],"type":"object","properties":{"path":{"title":"Path","type":"array","items":{"type":"string"}},"host":{"title":"Host","type":"array","items":{"type":"string"}},"query":{"title":"Query","type":"array","items":{}},"variable":{"title":"Variable","type":"array","items":{}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"Variable":{"title":"Variable","required":["id","type"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"type":{"title":"Type","type":"string"},"value":{"title":"Value","type":"string"}}},"mormo__schema__openapi_v3__Header":{"title":"Header","required":["schema"],"type":"object","properties":{"description":{"title":"Description","type":"string"},"required":{"title":"Required","type":"boolean"},"schema":{"$ref":"#/components/schemas/ParameterSchema"},"deprecated":{"title":"Deprecated","type":"boolean"},"allowEmptyValue":{"title":"Allowemptyvalue","type":"boolean"},"style":{"title":"Style","type":"string"},"explode":{"title":"Explode","type":"boolean"},"allowReserved":{"title":"Allowreserved","type":"boolean"},"example":{"title":"Example"},"examples":{"title":"Examples","type":"object","additionalProperties":{"anyOf":[{},{"$ref":"#/components/schemas/Reference"}]}},"content":{"title":"Content","type":"object","additionalProperties":{"$ref":"#/components/schemas/MediaType"}}}},"mormo__schema__openapi_v3__Info":{"title":"Info","type":"object","properties":{"description":{"title":"Description","type":"string"},"termsOfService":{"title":"Termsofservice","type":"string"},"title":{"title":"Title","type":"string"},"version":{"title":"Version","type":"string"},"contact":{"$ref":"#/components/schemas/Contact"},"license":{"$ref":"#/components/schemas/License"}}},"mormo__schema__openapi_v3__RequestBody":{"title":"RequestBody","required":["content"],"type":"object","properties":{"description":{"title":"Description","type":"string"},"content":{"title":"Content","type":"object","additionalProperties":{"$ref":"#/components/schemas/MediaType"}},"required":{"title":"Required","type":"boolean"}}},"mormo__schema__openapi_v3__Response":{"title":"Response","required":["description"],"type":"object","properties":{"description":{"title":"Description","type":"string"},"content":{"title":"Content","type":"object","additionalProperties":{"$ref":"#/components/schemas/MediaType"}},"headers":{"title":"Headers","type":"object","additionalProperties":{"anyOf":[{"$ref":"#/components/schemas/mormo__schema__openapi_v3__Header"},{"$ref":"#/components/schemas/Reference"}]}}}},"mormo__schema__openapi_v3__SaveDBResult":{"title":"SaveDBResult","required":["id","object"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"object":{"$ref":"#/components/schemas/OpenAPISchemaV3"}}},"mormo__schema__postman_collection_v2__Header":{"title":"Header","required":["key","value"],"type":"object","properties":{"key":{"title":"Key","type":"string"},"value":{"title":"Value","type":"string"}}},"mormo__schema__postman_collection_v2__Info":{"title":"Info","required":["name","_postman_id","schema","description"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"_postman_id":{"title":" Postman Id","type":"string"},"schema":{"title":"Schema","type":"string"},"description":{"$ref":"#/components/schemas/Description"}}},"mormo__schema__postman_collection_v2__RequestBody":{"title":"RequestBody","required":["mode","raw"],"type":"object","properties":{"mode":{"title":"Mode","enum":["raw","urlencoded","formdata","file","graphql"]},"raw":{"title":"Raw","type":"string"},"urlencoded":{"title":"Urlencoded","type":"array","items":{}},"formdata":{"title":"Formdata","type":"array","items":{}},"file":{"title":"File","type":"object"},"graphql":{"title":"Graphql","type":"object"},"diabled":{"title":"Diabled","type":"boolean","default":false}}},"mormo__schema__postman_collection_v2__Response":{"title":"Response","required":["id","name","originalRequest","status","code","header","body","cookie"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"originalRequest":{"$ref":"#/components/schemas/OriginalRequest"},"status":{"title":"Status","type":"string"},"code":{"title":"Code","type":"integer"},"header":{"title":"Header","type":"array","items":{"$ref":"#/components/schemas/mormo__schema__postman_collection_v2__Header"}},"body":{"title":"Body","type":"string"},"cookie":{"title":"Cookie","type":"array","items":{}}}},"mormo__schema__postman_collection_v2__SaveDBResult":{"title":"SaveDBResult","required":["id","object"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"object":{"$ref":"#/components/schemas/Collection"}}}}}}